% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolasso-package.R, R/fit-interface.R
\docType{package}
\name{bolasso}
\alias{bolasso}
\alias{bolasso-package}
\alias{bolasso.default}
\alias{bolasso.data.frame}
\alias{bolasso.matrix}
\alias{bolasso.formula}
\alias{bolasso.recipe}
\title{bolasso: A package for variable selection via bolasso algorithm}
\usage{
bolasso(x, ...)

\method{bolasso}{default}(x, ...)

\method{bolasso}{data.frame}(x, y, ...)

\method{bolasso}{matrix}(x, y, ...)

\method{bolasso}{formula}(formula, data, ...)

\method{bolasso}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{bolasso} object.
}
\description{
This package provides implementation of bolasso to select variables.
This is done via bootsrapping the training set and running a lasso in every bootstrap sample
Then, only variables that appear in most of the bootstrap samples are selected and a ridge
is trained with those variables

\code{bolasso()} fits a model.
}
\examples{
predictors <- mtcars[,c("mpg", "disp", "hp")]
outcome <- mtcars[, "vs"]

# XY interface
mod <- bolasso(predictors, outcome)

# Formula interface
mod2 <- bolasso(vs ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- bolasso(rec, mtcars)

}
\author{
\strong{Maintainer}: David Masip \email{david26694@gmail.com} (YOUR-ORCID-ID)

}
