% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-bridge.R
\name{bolasso_bridge}
\alias{bolasso_bridge}
\title{Fit bridge}
\usage{
bolasso_bridge(processed, ...)
}
\arguments{
\item{processed}{}
}
\description{
Fit bridge
}
\details{
The bridge connects the high level interface methods with the low level
implementation.

Pass through other objects to the implementation function as required.
}
\keyword{internal}
