% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-implementation.R
\name{bolasso_impl}
\alias{bolasso_impl}
\title{Low level bolasso implementation}
\usage{
bolasso_impl(predictors, outcome, n_bootsraps = 100,
  threshold_selection = 0.5, ...)
}
\arguments{
\item{predictors}{A numeric tibble of predictors.}

\item{outcome}{A tibble with outcome}
}
\description{
Low level bolasso implementation
}
\details{
\itemize{
\item x must be a matrix
\item y must be a vector
}
}
\keyword{internal}
