% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Extract indices used for each bootstrap replicate}
\usage{
bootstrap_samples(model)
}
\arguments{
\item{model}{A model of class \link{bolasso}.}
}
\value{
A named list of length equal to the number of bootstrap replicates
that contains the corresponding indices used to subsample the data.
}
\description{
This function extracts the indices used to bootstrap the data
for each replicate. This is helpful when doing some additional
analysis that requires knowledge of the exact bootstrap samples used.
}
