% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{ggplot.bolasso}
\alias{ggplot.bolasso}
\title{plot a \code{bolasso}}
\usage{
ggplot.bolasso(model, threshold_selection = NULL)
}
\arguments{
\item{model}{\code{bolasso} object to plot}

\item{threshold_selection}{If it is NULL, only variables with selection_rate above the one in the model
are plotted. We can set this value to lower or higher to plot more or less variables.}
}
\value{
A \code{ggplot} object, faceted and with the distribution of each variable in it.
}
\description{
The ggplot of a \code{bolasso} is a facet.
In each plot of the facet, there is an histogram of coefficients that the feature obtains in each
of the bootsrap samples. A red line shows the 0 coefficient value. If there are points at both sides
of the red line, it means that the feature has both positive and negative signs, which doesn't make sense.
}
\examples{

predictors <- mtcars[,c("mpg", "disp", "hp", "drat", "wt")]
outcome <- mtcars[, "vs"]


mod <- bolasso(predictors, outcome, n_bootstraps = 50)

ggplot(mod)
ggplot(mod, threshold_selection = 0.8)
ggplot(mod, threshold_selection = 0.2)

}
