% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_selection_thresholds}
\alias{plot_selection_thresholds}
\title{Plot each covariate's smallest variable selection threshold}
\usage{
plot_selection_thresholds(
  object = NULL,
  data = NULL,
  is_multinomial = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \link{bolasso} or \code{bolasso_fast}.
This argument is optional if you directly pass in the data via the
\code{data} argument. E.g. \code{data = selection_thresholds(object)}.}

\item{data}{A dataframe containing the selection thresholds. E.g.
obtained via \code{selection_thresholds(object)}. This argument is optional
if you directly pass a \code{bolasso} or \code{bolasso_fast} object via the \code{object}
argument.}

\item{is_multinomial}{A boolean indicating if the data provided is from a
multinomial regression model. Defaults to FALSE. This only needs to be
provided if \code{data} is passed in directly. Otherwise this will be inferred
from the model object.}

\item{...}{Additional arguments to pass directly to
\link{selection_thresholds}.}
}
\value{
A \code{ggplot} object
}
\description{
Plot the results of the \link{selection_thresholds} function.
}
\seealso{
\code{\link[=selection_thresholds]{selection_thresholds()}}
}
