% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{summary.bolasso}
\alias{summary.bolasso}
\title{Summarise a \code{bolasso}}
\usage{
\method{summary}{bolasso}(model)
}
\arguments{
\item{model}{\code{bolasso} object to summarise}
}
\value{
A \code{tbl}.
}
\description{
The summary of a bolasso is structured as a \code{tbl}.
Every row of the \code{tbl} represents a feature and it always has 4 columns:
* feature: name of the feature
* selection_ratio: proportion of the times that the feature is selected
* selected_bolasso: TRUE if selected_ratio is large enough for the feature to be selected
* coefficient: Coefficient in ridge regression trained afterwards
}
\examples{

predictors <- mtcars[,c("mpg", "disp", "hp", "drat", "wt")]
outcome <- mtcars[, "vs"]


mod <- bolasso(predictors, outcome, n_bootsraps = 10)

summary(mod)

}
