% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactions}
\alias{transactions}
\title{Customer transaction data}
\format{
Dataframe with columns
\describe{
\item{target}{An integer indicating whether a customer engaged in a transaction.}
\item{var_i}{200 numeric features of various customer characteristics.}
}
}
\usage{
transactions
}
\description{
Predict whether customers will make a specific transaction based on
a rich set of user features.
}
\keyword{datasets}
