% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-implementation.R
\name{bolasso_impl}
\alias{bolasso_impl}
\title{Low level bolasso implementation}
\usage{
bolasso_impl(predictors, outcome, n_bootstraps = 100,
  threshold_selection = 0.5, ...)
}
\arguments{
\item{predictors}{A numeric tibble of predictors.}

\item{outcome}{A tibble with outcome}
}
\description{
Low level bolasso implementation
}
\details{
First, we use rsample to generate bootstrap samples.
Then, bolasso for variable selection.
And finally, ridge regression on the selected variables.
The selected variavbles are the ones appearing more than threshold_selection times and
that have consistent signs (that is, when selected, the sign of the coefficient is always
the same).
}
\keyword{internal}
